fixed listing css

+++ auctions/models.py
@@ -9,7 +9,7 @@ class Category(models.Model):
         return self.name
 class Listing(models.Model):
     title = models.CharField(max_length = 100)
-    start_bid = models.IntegerField()
+    start_bid = models.DecimalField(decimal_places=2, max_digits = 1000000)
     url = models.CharField(max_length = 200, blank = True)
     description = models.CharField(max_length = 200, default = "")
     category = models.ForeignKey(Category, on_delete = models.CASCADE, related_name = "category_listing", default  = "")
